rename hostyarg to yarg
