+++ .gitignore
@@ -1,4 +1,5 @@
-hostyarg/hostyarg
+hostyarg/cmd/yarg/yarg
+hostyarg/yarg
 *.uf2
 .vscode
 build
