+++ docs/release_README
@@ -32,13 +32,13 @@ and the built in led should illuminate
 
 extras/yarg-lang-0.3.0.vsix is a vscode syntax highligher for yarg. Install it via the 'Extensions: Install from visx...' command.
 
-*** hostyarg
+*** Host yarg
 
-Use go install github.com/yarg-lang/yarg-lang/hostyarg@0.3.0 to install hostyarg.
+Use go install github.com/yarg-lang/yarg-lang/hostyarg/cmd/yarg@0.3.0 to install `yarg`.
 
 You can use this to add yarg files to your image, for your own development, eg:
 
-$ hostyarg addfile -fs yarg-lang-pico-0.3.0.uf2 -add my_yarg.ya
+$ yarg addfile -fs yarg-lang-pico-0.3.0.uf2 -add my_yarg.ya
 
 install the uf2 as above, and after the pico reboots, at the yarg prompt:
