+++ hostyarg/hostyarg_test.go
@@ -1,4 +1,4 @@
-package main
+package hostyarg
 
 import (
 	"flag"
