+++ test/README.md
@@ -2,7 +2,7 @@
 
 test/ contains test suites:
 
-`hostyarg/` - the primary Yarg Language test suite, hosted by hostyarg.
+`hostyarg/` - the primary Yarg Language test suite, hosted by `yarg`.
 
 `bc/` - a 'bc' based test suite of the cyarg/big-int integer library.
 
@@ -10,7 +10,7 @@ test/ contains test suites:
 
 ## hostyarg Yarg Language
 
-A test suite that exercises the Yarg Language. Each file is parsed twice, once by hostyarg for expected results, and then by cyarg itself. 
+A test suite that exercises the Yarg Language. Each file is parsed twice, once by `yarg` for expected results, and then by `cyarg` itself. 
 
 See [yarg_test.md](../docs/yarg_test.md).
