+++ tools/build-specimen.sh
@@ -5,6 +5,6 @@ then
     rm build/specimen-fs.uf2
 fi
 
-./bin/hostyarg format -fs build/specimen-fs.uf2
+./bin/yarg format -fs build/specimen-fs.uf2
 ./tools/add-yarg-stdlib.sh build/specimen-fs.uf2
 ./tools/add-specimen.sh build/specimen-fs.uf2
